requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0